# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/houses": {
      "get": {
        "operationId": "HousesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Houses"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/houses/{id}/members": {
      "get": {
        "operationId": "HousesController_findMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseMemberResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Houses"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/houses/members/{user_id}": {
      "patch": {
        "operationId": "HousesController_assignHouse",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignHouseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignHouseResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Houses"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/cycles": {
      "post": {
        "operationId": "GamificationCyclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCycleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationCycles"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GamificationCyclesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CycleResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "GamificationCycles"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/cycles/{id}/close": {
      "patch": {
        "operationId": "GamificationCyclesController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationCycles"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/cycles/active": {
      "get": {
        "operationId": "GamificationCyclesController_getActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationCycles"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/tasks": {
      "post": {
        "operationId": "GamificationTasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationTasks"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GamificationTasksController_list",
        "parameters": [
          {
            "name": "include_inactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "GamificationTasks"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/tasks/{id}": {
      "patch": {
        "operationId": "GamificationTasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationTasks"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/submissions": {
      "post": {
        "operationId": "GamificationSubmissionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationSubmissions"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "GamificationSubmissionsController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "GamificationSubmissions"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/submissions/{id}/review": {
      "patch": {
        "operationId": "GamificationSubmissionsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "GamificationSubmissions"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/leaderboard/podium": {
      "get": {
        "operationId": "GamificationLeaderboardController_getPodium",
        "parameters": [
          {
            "name": "house_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycle_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PodiumEntryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "GamificationLeaderboard"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/gamification/leaderboard": {
      "get": {
        "operationId": "GamificationLeaderboardController_getLeaderboard",
        "parameters": [
          {
            "name": "cycle_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "GamificationLeaderboard"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WattAPI",
    "description": "Documentação gerada a partir dos schemas zod de request/response e do metadata de @RoutePolicy de cada rota.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "HouseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "total_points": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "total_points"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "name": "Lumina",
          "total_points": 150
        }
      },
      "HouseMemberResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "house_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "sector",
          "house_id"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "name": "Ana Souza",
          "email": "ana@empresa.com",
          "role": "consultor",
          "sector": "projetos",
          "house_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
      "AssignHouseDto": {
        "type": "object",
        "properties": {
          "house_id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
            "nullable": true
          }
        },
        "required": [
          "house_id"
        ],
        "example": {
          "house_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
      "AssignHouseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sector": {
            "type": "string"
          },
          "house_id": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "sector",
          "house_id"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "name": "Ana Souza",
          "email": "ana@empresa.com",
          "role": "consultor",
          "sector": "projetos",
          "house_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
      "CreateCycleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ],
        "example": {
          "name": "1º Semestre 2026"
        }
      },
      "CycleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "started_at": {
            "type": "string"
          },
          "ended_at": {
            "type": "string",
            "nullable": true
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "started_at",
          "ended_at",
          "created_by",
          "created_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "name": "1º Semestre 2026",
          "started_at": "2026-06-04T00:00:00Z",
          "ended_at": null,
          "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "created_at": "2026-06-04T00:00:00Z"
        }
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "points": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0
          }
        },
        "required": [
          "title",
          "description",
          "points"
        ],
        "example": {
          "title": "Participar de vídeo do marketing",
          "description": "Aparecer em um vídeo oficial da Watt Consultoria",
          "points": 50
        }
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "points",
          "is_active",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "title": "Participar de vídeo do marketing",
          "description": "Aparecer em um vídeo oficial da Watt Consultoria",
          "points": 50,
          "is_active": true,
          "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "created_at": "2026-06-04T00:00:00.000Z",
          "updated_at": "2026-06-04T00:00:00.000Z"
        }
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "points": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          }
        },
        "example": {
          "points": 75,
          "is_active": false
        }
      },
      "CreateSubmissionDto": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "file_path": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "task_id",
          "description",
          "file_path"
        ],
        "example": {
          "task_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "description": "Participei do vídeo em 04/06/2026",
          "file_path": "proofs/uuid-do-usuario/nome-do-arquivo.pdf"
        }
      },
      "SubmissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "house_id": {
            "type": "string"
          },
          "cycle_id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "file_url": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "rejection_reason": {
            "type": "string",
            "nullable": true
          },
          "reviewed_by": {
            "type": "string",
            "nullable": true
          },
          "reviewed_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "user_id",
          "house_id",
          "cycle_id",
          "description",
          "file_url",
          "status",
          "rejection_reason",
          "reviewed_by",
          "reviewed_at",
          "created_at",
          "updated_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "task_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "user_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "house_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "cycle_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "description": "Participei do vídeo em 04/06/2026",
          "file_url": "https://storage.example.com/gamification-proofs/signed-url",
          "status": "pending",
          "rejection_reason": null,
          "reviewed_by": null,
          "reviewed_at": null,
          "created_at": "2026-06-04T00:00:00.000Z",
          "updated_at": "2026-06-04T00:00:00.000Z"
        }
      },
      "ReviewSubmissionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "rejection_reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "example": {
          "status": "approved",
          "rejection_reason": "opcional — apenas para rejections"
        }
      },
      "PodiumEntryDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          },
          "points_contributed": {
            "type": "number"
          },
          "approved_count": {
            "type": "number"
          }
        },
        "required": [
          "user_id",
          "user_name",
          "points_contributed",
          "approved_count"
        ],
        "example": {
          "user_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "user_name": "Danilo Silva",
          "points_contributed": 150,
          "approved_count": 3
        }
      },
      "LeaderboardEntryDto": {
        "type": "object",
        "properties": {
          "house_id": {
            "type": "string"
          },
          "house_name": {
            "type": "string"
          },
          "total_points": {
            "type": "number"
          }
        },
        "required": [
          "house_id",
          "house_name",
          "total_points"
        ],
        "example": {
          "house_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "house_name": "Nexus",
          "total_points": 210
        }
      },
...
    }
  }
}